+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Option to disable the typing animation of indicators in text fields
 - Exact string search for GitHub resource
 - Investigate option in context menu to send indicators to the popup
+- Option in the context menu to add selected text as a node to the graph
 - Icon to mark the tools that support the automatic graph population feature
 - Icon to mark the tools that require user interaction
 - Icon to mark the tools that require an account